requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Database
aiosqlite==0.20.0

# LLM
google-generativeai==0.8.3

# Data Validation
pydantic==2.9.2

# CORS
python-multipart==0.0.12

# Utilities
python-dateutil==2.9.0
